TOT-5 : [CHORE] 도메인별 패키지 분리 및 적용
