--- backend/src/main/java/com/triportreat/backend/entity/Region.java
@@ -1,37 +1,0 @@
-package com.triportreat.backend.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Getter
-public class Region {
-
-    @Id
-    private Long id;
-
-    @Column(nullable = false)
-    private String name;
-
-    private String imageOrigin;
-
-    private String imageThumbnail;
-
-    @Column(nullable = false)
-    private Double latitude;
-
-    @Column(nullable = false)
-    private Double longitude;
-
-    @Column(length = 65535)
-    private String overview;
-
-}
