+++ backend/src/main/java/com/triportreat/backend/plan/entity/SchedulePlace.java
@@ -1,5 +1,6 @@
-package com.triportreat.backend.entity;
+package com.triportreat.backend.plan.entity;
 
+import com.triportreat.backend.place.entity.Place;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -7,7 +8,6 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import java.time.LocalDate;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -18,16 +18,26 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 @Getter
-public class Schedule {
+public class SchedulePlace {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne
-    @JoinColumn(name = "plan_id")
-    private Plan plan;
+    @JoinColumn(name = "schedule_id")
+    private Schedule schedule;
+
+    @ManyToOne
+    @JoinColumn(name = "place_id")
+    private Place place;
 
     @Column(nullable = false)
-    private LocalDate visitDate;
+    private Integer visitOrder;
+
+    //null이어도 합계 계산할 때 문제없는가?
+    private Integer expense;
+
+    @Column(length = 65535)
+    private String memo;
 
 }
