+++ backend/src/main/java/com/triportreat/backend/user/entity/User.java
@@ -1,4 +1,34 @@
 package com.triportreat.backend.user.entity;
 
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@Getter
 public class User {
+
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(length = 10)
+    private String name;
+
+    @Column(length = 10, nullable = false)
+    private String nickname;
+
+    @Column(length = 30)
+    private String email;
+
+    private String imageOrigin;
+
 }
